TV: Move SigningInViewModel to its own file
